Lint .vue files using eslint-plugin-html
Initially attempted to integrate eslint-plugin-vue, but couldn't get
this working with standardjs (it seems to be geared towards running
eslint directly with additional configuration).

May revisit switching to eslint in the future, but for now the html
plugin seems to do a reasonable job (parsing the code within <script>
tags).