Add search component using the Google Maps API
Currently only works when the map is visible since it relies on the
marker moving to trigger a map update and consequently query for new
events. Next step will be to refactor this to work when searching on
the list view as well.

Adds a new environment configuration value for setting the Google Maps
API key.