Refactoring of events compoonents
A step towards having the search component work on both the map and the
list views. Pushes more control up to the Events component to make it
simpler to share logic between the two views (map and list)

*Note:* Search component only works on the map (and not the list view)
as it depends on the map bounds updating to requery for events.

* Update Google API key comment to specify that it also requires access
  to the Places API
* Have the Events component manage querying for the events and injecting
  them rather than the two display components needing to know how to
  query
* Have Events component manage opening an event