Store zoom and center instead of bounds in URL
Rather than storing the bounds in the URL, store the map center and zoom
level and pass those to the EventMap component.

This will be used to calculate the bounds of the map to allow querying
by address for when either the EventMap or the EventList is shown.

Currently still relying on the map bounds changing to trigger querying.