Calculate bounding box polygon when searching by address
Perform similar calculations to what LeafletJS is probably doing to get
the bounding box of the map based on the size of the router view in
Events. This allows for the appropriate bounding box to be calculated
when searching on the list page.

Previously considered was continuing to store the bounds in the URL but
we found that when switching from the list view to the map view, the map
would actually zoom out one level in its attempts to fit the bounds. By
relying on the center and zoom solely, we avoid this issue.