Add map locate control
Add locate.locatecontrol to the map to allow the user to locate
themselves. Useful if the user has panned and wants to return to their
location.

This changes the geolocation behavior on the start page to, instead of
doing the location and loading the map, simply load the map and ask the
map to locate the user (via the added locatecontrol). Though we wanted
to preserve the behavior of location before map load, I wasn't able to
get this to play nicely with rendering the user's location via the newly
added locate control which doesn't have an interface for setting the
current location (it depends on receiving a Leaflet.LocateEvent).
I Opened https://github.com/domoritz/leaflet-locatecontrol/issues/209 to
get the maintainer's feedback on adding an interface to set the location.